Make imagec work as a library
