+++ lib/apiservers/engine/backends/cache/image_cache.go
@@ -16,7 +16,6 @@ package cache
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"os"
@@ -34,15 +33,6 @@ import (
 	"github.com/vmware/vic/lib/apiservers/portlayer/models"
 	"github.com/vmware/vic/lib/guest"
 	"github.com/vmware/vic/lib/metadata"
-)
-
-var (
-	// ErrCacheNotUpdated is returned when an operation is attempted while the cache has
-	// not yet hydrated.
-	ErrCacheNotUpdated = errors.New("Image cache not updated")
-
-	// CacheNotUpdated indicates whether or not the cache has been successfully hydrated
-	CacheNotUpdated = true
 )
 
 // ImageCache is an in-memory cache of image metadata. It is refreshed at startup
@@ -66,10 +56,6 @@ func NewImageCache() *ImageCache {
 
 // Update adds new images to the cache
 func (c *ImageCache) Update(client *client.PortLayer) error {
-	c.m.Lock()
-	defer c.m.Unlock()
-
-	CacheNotUpdated = true
 
 	log.Debugf("Updating image cache...")
 
@@ -111,55 +97,17 @@ func (c *ImageCache) Update(client *client.PortLayer) error {
 		}
 
 		if imageConfig.ImageID != "" {
-			var imageID string
-
-			// Don't assume the image id in image has "sha256:<id> as format.  We store it in
-			// this fomat to make it easier to lookup by digest
-			if strings.HasPrefix(imageConfig.ImageID, "sha") {
-				imageID = imageConfig.ImageID
-			} else {
-				imageID = "sha256:" + imageConfig.ImageID
-			}
-
-			c.cacheByID[imageID] = imageConfig
-
-			// Normalize the name stored in imageConfig using Docker's reference code
-			ref, err := reference.WithName(imageConfig.Name)
-			if err != nil {
-				log.Errorf("Tried to create reference from %s: %s", imageConfig.Name, err.Error())
-				continue
-			}
-
-			for id := range imageConfig.Tags {
-				tag := imageConfig.Tags[id]
-				ref, err = reference.WithTag(ref, tag)
-				if err != nil {
-					log.Errorf("Tried to create tagged reference from %s and tag %s: %s", imageConfig.Name, tag, err.Error())
-					continue
-				}
-
-				if tagged, ok := ref.(reference.NamedTagged); ok {
-					taggedName := fmt.Sprintf("%s:%s", tagged.Name(), tagged.Tag())
-					c.cacheByName[taggedName] = imageConfig
-				} else {
-					c.cacheByName[ref.Name()] = imageConfig
-				}
-			}
+			c.AddImage(imageConfig)
 		}
 	}
 
-	CacheNotUpdated = false
 	return nil
 }
 
 // GetImages returns a slice containing metadata for all cached images
-func (c *ImageCache) GetImages() ([]*metadata.ImageConfig, error) {
+func (c *ImageCache) GetImages() []*metadata.ImageConfig {
 	c.m.RLock()
 	defer c.m.RUnlock()
-
-	if CacheNotUpdated {
-		return nil, ErrCacheNotUpdated
-	}
 
 	result := make([]*metadata.ImageConfig, 0, len(c.cacheByID))
 	for _, image := range c.cacheByID {
@@ -168,29 +116,20 @@ func (c *ImageCache) GetImages() ([]*metadata.ImageConfig, error) {
 		result = append(result, newImage)
 	}
 
-	return result, nil
+	return result
 }
 
-func (c *ImageCache) GetImageByDigest(digest digest.Digest) (*metadata.ImageConfig, error) {
+func (c *ImageCache) GetImageByDigest(digest digest.Digest) *metadata.ImageConfig {
 	c.m.RLock()
 	defer c.m.RUnlock()
 
-	if CacheNotUpdated {
-		return nil, ErrCacheNotUpdated
-	}
-
-	config := c.cacheByID[string(digest)]
-	return config, nil
+	return c.cacheByID[string(digest)]
 }
 
 // Looks up image by reference.Named
-func (c *ImageCache) GetImageByNamed(named reference.Named) (*metadata.ImageConfig, error) {
+func (c *ImageCache) GetImageByNamed(named reference.Named) *metadata.ImageConfig {
 	c.m.RLock()
 	defer c.m.RUnlock()
-
-	if CacheNotUpdated {
-		return nil, ErrCacheNotUpdated
-	}
 
 	var config *metadata.ImageConfig
 
@@ -207,5 +146,45 @@ func (c *ImageCache) GetImageByNamed(named reference.Named) (*metadata.ImageConf
 		}
 	}
 
-	return config, nil
+	return config
+}
+
+// AddImage adds an image to the image cache
+func (c *ImageCache) AddImage(imageConfig *metadata.ImageConfig) {
+	c.m.Lock()
+	defer c.m.Unlock()
+	var imageID string
+
+	// Don't assume the image id in image has "sha256:<id> as format.  We store it in
+	// this fomat to make it easier to lookup by digest
+	if strings.HasPrefix(imageConfig.ImageID, "sha") {
+		imageID = imageConfig.ImageID
+	} else {
+		imageID = "sha256:" + imageConfig.ImageID
+	}
+
+	c.cacheByID[imageID] = imageConfig
+
+	// Normalize the name stored in imageConfig using Docker's reference code
+	ref, err := reference.WithName(imageConfig.Name)
+	if err != nil {
+		log.Errorf("Tried to create reference from %s: %s", imageConfig.Name, err.Error())
+		return
+	}
+
+	for id := range imageConfig.Tags {
+		tag := imageConfig.Tags[id]
+		ref, err = reference.WithTag(ref, tag)
+		if err != nil {
+			log.Errorf("Tried to create tagged reference from %s and tag %s: %s", imageConfig.Name, tag, err.Error())
+			return
+		}
+
+		if tagged, ok := ref.(reference.NamedTagged); ok {
+			taggedName := fmt.Sprintf("%s:%s", tagged.Name(), tagged.Tag())
+			c.cacheByName[taggedName] = imageConfig
+		} else {
+			c.cacheByName[ref.Name()] = imageConfig
+		}
+	}
 }
