+++ lib/imagec/docker.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package imagec
 
 import (
 	"archive/tar"
@@ -63,11 +63,11 @@ type Manifest struct {
 }
 
 // LearnRegistryURL returns the registry URL after making sure that it responds to queries
-func LearnRegistryURL(options ImageCOptions) (string, error) {
-	defer trace.End(trace.Begin(options.registry))
+func LearnRegistryURL(options Options) (string, error) {
+	defer trace.End(trace.Begin(options.Registry))
 
 	req := func(schema string) (string, error) {
-		registry := fmt.Sprintf("%s://%s/v2/", schema, options.registry)
+		registry := fmt.Sprintf("%s://%s/v2/", schema, options.Registry)
 
 		url, err := url.Parse(registry)
 		if err != nil {
@@ -76,15 +76,16 @@ func LearnRegistryURL(options ImageCOptions) (string, error) {
 		log.Debugf("URL: %s", url)
 
 		fetcher := NewURLFetcher(FetcherOptions{
-			Timeout:            options.timeout,
-			Username:           options.username,
-			Password:           options.password,
-			InsecureSkipVerify: options.insecureSkipVerify,
+			Timeout:            options.Timeout,
+			Username:           options.Username,
+			Password:           options.Password,
+			InsecureSkipVerify: options.InsecureSkipVerify,
 		})
 		headers, err := fetcher.Head(url)
 		if err != nil {
 			return "", err
 		}
+		log.Info("Successfully ran fetcher.Head()")
 		// v2 API requires this check
 		if headers.Get("Docker-Distribution-API-Version") != "registry/2.0" {
 			return "", fmt.Errorf("Missing Docker-Distribution-API-Version header")
@@ -95,7 +96,7 @@ func LearnRegistryURL(options ImageCOptions) (string, error) {
 	// first try https
 	log.Debugf("Trying https scheme")
 	registry, err := req("https")
-	if err != nil && options.insecureAllowHTTP {
+	if err != nil && options.InsecureAllowHTTP {
 		// fallback to http if it's allowed
 		log.Debugf("Falling back to http scheme")
 		registry, err = req("http")
@@ -105,22 +106,22 @@ func LearnRegistryURL(options ImageCOptions) (string, error) {
 }
 
 // LearnAuthURL returns the URL of the OAuth endpoint
-func LearnAuthURL(options ImageCOptions) (*url.URL, error) {
-	defer trace.End(trace.Begin(options.image + "/" + options.tag))
+func LearnAuthURL(options Options) (*url.URL, error) {
+	defer trace.End(trace.Begin(options.Image + "/" + options.Tag))
 
-	url, err := url.Parse(options.registry)
+	url, err := url.Parse(options.Registry)
 	if err != nil {
 		return nil, err
 	}
-	url.Path = path.Join(url.Path, options.image, "manifests", options.tag)
+	url.Path = path.Join(url.Path, options.Image, "manifests", options.Tag)
 
 	log.Debugf("URL: %s", url)
 
 	fetcher := NewURLFetcher(FetcherOptions{
-		Timeout:            options.timeout,
-		Username:           options.username,
-		Password:           options.password,
-		InsecureSkipVerify: options.insecureSkipVerify,
+		Timeout:            options.Timeout,
+		Username:           options.Username,
+		Password:           options.Password,
+		InsecureSkipVerify: options.InsecureSkipVerify,
 	})
 	// We expect docker registry to return a 401 to us - with a WWW-Authenticate header
 	// We parse that header and learn the OAuth endpoint to fetch OAuth token.
@@ -131,14 +132,14 @@ func LearnAuthURL(options ImageCOptions) (*url.URL, error) {
 
 	// Private registry returned the manifest directly as auth option is optional.
 	// https://github.com/docker/distribution/blob/master/docs/configuration.md#auth
-	if err == nil && options.registry != DefaultDockerURL && fetcher.IsStatusOK() {
+	if err == nil && options.Registry != DefaultDockerURL && fetcher.IsStatusOK() {
 		log.Debugf("%s does not support OAuth", url)
 		return nil, nil
 	}
 
 	// Do we even have the image on that registry
 	if err != nil && fetcher.IsStatusNotFound() {
-		return nil, fmt.Errorf("%s:%s does not exists at %s", options.image, options.tag, options.registry)
+		return nil, fmt.Errorf("%s:%s does not exists at %s", options.Image, options.Tag, options.Registry)
 	}
 
 	return nil, fmt.Errorf("%s returned an unexpected response: %s", url, err)
@@ -151,10 +152,10 @@ func FetchToken(url *url.URL) (*Token, error) {
 	log.Debugf("URL: %s", url)
 
 	fetcher := NewURLFetcher(FetcherOptions{
-		Timeout:            options.timeout,
-		Username:           options.username,
-		Password:           options.password,
-		InsecureSkipVerify: options.insecureSkipVerify,
+		Timeout:            options.Timeout,
+		Username:           options.Username,
+		Password:           options.Password,
+		InsecureSkipVerify: options.InsecureSkipVerify,
 	})
 	tokenFileName, err := fetcher.Fetch(url)
 	if err != nil {
@@ -187,30 +188,30 @@ func FetchToken(url *url.URL) (*Token, error) {
 }
 
 // FetchImageBlob fetches the image blob
-func FetchImageBlob(options ImageCOptions, image *ImageWithMeta) (string, error) {
-	defer trace.End(trace.Begin(options.image + "/" + image.layer.BlobSum))
+func FetchImageBlob(options Options, image *ImageWithMeta) (string, error) {
+	defer trace.End(trace.Begin(options.Image + "/" + image.layer.BlobSum))
 
 	id := image.ID
 	layer := image.layer.BlobSum
 	meta := image.meta
 	diffID := ""
 
-	url, err := url.Parse(options.registry)
+	url, err := url.Parse(options.Registry)
 	if err != nil {
 		return diffID, err
 	}
-	url.Path = path.Join(url.Path, options.image, "blobs", layer)
+	url.Path = path.Join(url.Path, options.Image, "blobs", layer)
 
 	log.Debugf("URL: %s\n ", url)
 
 	progress.Update(po, image.String(), "Pulling fs layer")
 
 	fetcher := NewURLFetcher(FetcherOptions{
-		Timeout:            options.timeout,
-		Username:           options.username,
-		Password:           options.password,
-		Token:              options.token,
-		InsecureSkipVerify: options.insecureSkipVerify,
+		Timeout:            options.Timeout,
+		Username:           options.Username,
+		Password:           options.Password,
+		Token:              options.Token,
+		InsecureSkipVerify: options.InsecureSkipVerify,
 	})
 	imageFileName, err := fetcher.FetchWithProgress(url, image.String())
 	if err != nil {
@@ -318,23 +319,23 @@ func FetchImageBlob(options ImageCOptions, image *ImageWithMeta) (string, error)
 }
 
 // FetchImageManifest fetches the image manifest file
-func FetchImageManifest(options ImageCOptions) (*Manifest, error) {
-	defer trace.End(trace.Begin(options.image + "/" + options.tag))
+func FetchImageManifest(options Options) (*Manifest, error) {
+	defer trace.End(trace.Begin(options.Image + "/" + options.Tag))
 
-	url, err := url.Parse(options.registry)
+	url, err := url.Parse(options.Registry)
 	if err != nil {
 		return nil, err
 	}
-	url.Path = path.Join(url.Path, options.image, "manifests", options.tag)
+	url.Path = path.Join(url.Path, options.Image, "manifests", options.Tag)
 
 	log.Debugf("URL: %s", url)
 
 	fetcher := NewURLFetcher(FetcherOptions{
 		Timeout:            10 * time.Second,
-		Username:           options.username,
-		Password:           options.password,
-		Token:              options.token,
-		InsecureSkipVerify: options.insecureSkipVerify,
+		Username:           options.Username,
+		Password:           options.Password,
+		Token:              options.Token,
+		InsecureSkipVerify: options.InsecureSkipVerify,
 	})
 	manifestFileName, err := fetcher.Fetch(url)
 	if err != nil {
@@ -361,12 +362,12 @@ func FetchImageManifest(options ImageCOptions) (*Manifest, error) {
 		return nil, err
 	}
 
-	if manifest.Name != options.image {
-		return nil, fmt.Errorf("name doesn't match what was requested, expected: %s, downloaded: %s", options.image, manifest.Name)
+	if manifest.Name != options.Image {
+		return nil, fmt.Errorf("name doesn't match what was requested, expected: %s, downloaded: %s", options.Image, manifest.Name)
 	}
 
-	if manifest.Tag != options.tag {
-		return nil, fmt.Errorf("tag doesn't match what was requested, expected: %s, downloaded: %s", options.tag, manifest.Tag)
+	if manifest.Tag != options.Tag {
+		return nil, fmt.Errorf("tag doesn't match what was requested, expected: %s, downloaded: %s", options.Tag, manifest.Tag)
 	}
 
 	digest, err := getManifestDigest(content)
